full crud on wishlist gql, added basic styles and tweaked some reslvers to hadnle said crud

+++ server/schema/lib/wishlist_resolvers.js
@@ -1,30 +1,65 @@
-const Wishlist = require('../../models/Wishlist');
+const User = require("../../models/User");
+const Wishlist = require("../../models/Wishlist");
 
 const wishlist_resolvers = {
-    Query: {
-        
+  Query: {
+    async getWishlists(_, __, context) {
+      if (!context || !context.user || !context.user._id)
+        throw new Error("User not logged in");
+
+      try {
+        const wishlists = await Wishlist.find({ user: context.user._id });
+
+        return wishlists;
+      } catch (error) {
+        console.error("error fetching wishlists");
+      }
+    },
+  },
+
+  Mutation: {
+    async createWishlist(_, { name }, context) {
+      if (!context || !context.user || !context.user._id)
+        throw new Error("User is not authenticated");
+      const wishlist = await Wishlist.create({
+        name,
+        user: context.user._id,
+      });
+      await User.findByIdAndUpdate(
+        context.user._id,
+        { $push: { wishlists: wishlist._id } },
+        { new: true }
+      );
+      return wishlist;
     },
 
-    Mutation: {
-        async createWishlist(_, args) {
-            const wishlist = await Wishlist.create(args);
+    async updateWishlist(_, { id, name }, context) {
+      if (!context || !context.user || !context.user._id)
+        throw new Error("User is not authenticated");
+      const wishlist = await Wishlist.findByIdAndUpdate(
+        id,
+        { name },
+        { new: true }
+      );
 
-            return wishlist;
-        },
+      return wishlist;
+    },
 
-        async updateWishlist(_, { id, ...args }) {
-            const wishlist = await Wishlist.findByIdAndUpdate(id, args, { new: true });
+    async deleteWishlist(_, { id }, context) {
+      if (!context || !context.user || !context.user._id)
+        throw new Error("User is not authenticated");
+      const wishlist = await Wishlist.findByIdAndDelete(id);
+      if (!wishlist) throw new Error("no wishlist found");
 
-            return wishlist;
-        },
+      await User.findByIdAndUpdate(
+        context.user._id,
+        { $pull: { wishlists: id } },
+        { new: true }
+      );
 
-        async deleteWishlist(_, { id }) {
-            const wishlist = await Wishlist.findByIdAndDelete(id);
-
-            return wishlist;
-        }
-    }
-}
-
+      return wishlist;
+    },
+  },
+};
 
 module.exports = wishlist_resolvers;
